Example: clean up
- use provider
- use connectivy plus
- move code snippet button and example into their own files
