+++ example/pubspec.yaml
@@ -8,12 +8,14 @@ environment:
   flutter: ">=3.3.0"
 
 dependencies:
+  connectivity_plus: ^3.0.2
   flutter:
     sdk: flutter
   flutter_highlight: ^0.7.0
   handy_window: ^0.1.2
   http: ^0.13.5
   provider: ^6.0.2
+  safe_change_notifier: ^0.2.0
   yaru: ^0.4.3
   yaru_icons: ^0.2.6
   yaru_widgets:
